pyproject.toml
[tool.poetry]
name = "fut_sniping_bot"
version = "0.1.0"
description = "FIFA FUT Sniping Bot"
authors = ["Todd <toddbirchard@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/toddbirchard/fut_sniping_bot/"
repository = "https://github.com/toddbirchard/fut_sniping_bot/"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "*"
loguru = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
mock = "*"
mypy = "*"
black = "^20.8b1"
isort = "*"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
